updates bap-toolkit tests
Adds tests for bap-toolkit analysis.
We add a set of toy programs, one toy per analysis. Every toy program
contains a bad case and (optionally) a good one. The former one must
be discovered during the analysis, and a latter one - must be not.
We don't pretend these tests are supposed to cover 100% of our needs
in testing, but we need them anyway to track bugs faster, prevent
regressions in our codebase and etc.
Also, there is a set of proven, correct data (incidents) that are
used as a around truth in our test system. Any declination from it
is considered as a FAIL case, i.e. we don't allow to have any
false positivites for toy programs.

Also, previous test we removed. Althoug the idea is good, the
implementation looks unusable. We'll return to them later.